Used count instead of first to check existence of record.

+++ service/core/action/medium/delete.go
@@ -52,6 +52,17 @@ func delete(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	var cnt int
+	config.DB.Model(&model.Post{}).Where(&model.Post{
+		SpaceID:          uint(sID),
+		FeaturedMediumID: uint(id),
+	}).Count(&cnt)
+
+	if cnt != 0 {
+		errorx.Render(w, errorx.Parser(util.CannotDeleteError()))
+		return
+	}
+
 	config.DB.Delete(&result)
 
 	renderx.JSON(w, http.StatusOK, nil)
